Add gitlab_runner e2e
