+++ gitlab_runner/tests/test_integration.py
@@ -6,36 +6,10 @@ import pytest
 from requests.exceptions import ConnectionError
 
 from datadog_checks.gitlab_runner import GitlabRunnerCheck
-
-from .common import ALLOWED_METRICS, BAD_CONFIG, CONFIG, CUSTOM_TAGS, GITLAB_RUNNER_TAGS, HOST
-
-
-def test_check(aggregator):
-    """
-    Basic Test for gitlab integration.
-    """
-    gitlab_runner = GitlabRunnerCheck('gitlab_runner', CONFIG['init_config'], {}, instances=CONFIG['instances'])
-
-    gitlab_runner.check(CONFIG['instances'][0])
-
-    aggregator.assert_service_check(
-        GitlabRunnerCheck.MASTER_SERVICE_CHECK_NAME,
-        status=GitlabRunnerCheck.OK,
-        tags=GITLAB_RUNNER_TAGS + CUSTOM_TAGS,
-        count=1,
-    )
-
-    aggregator.assert_service_check(
-        GitlabRunnerCheck.PROMETHEUS_SERVICE_CHECK_NAME, status=GitlabRunnerCheck.OK, tags=CUSTOM_TAGS, count=1
-    )
-
-    for metric in ALLOWED_METRICS:
-        if metric.startswith('ci_runner'):
-            aggregator.assert_metric("gitlab_runner.{}".format(metric))
-        else:
-            aggregator.assert_metric("gitlab_runner.{}".format(metric), tags=CUSTOM_TAGS, count=1)
+from .common import BAD_CONFIG, CUSTOM_TAGS, HOST
 
 
+@pytest.mark.usefixtures("dd_environment")
 def test_connection_failure(aggregator):
     """
     Make sure we're failing when the URL isn't right
