+++ gitlab_runner/tox.ini
@@ -5,6 +5,8 @@ envlist =
     py{27,37}-10.8.0
 
 [testenv]
+description=
+    py{27,37}: e2e ready
 dd_check_style = true
 usedevelop = true
 platform = linux|darwin|win32
